Do not auto-enable the cds remote (#20)
It seems more likely than not that people don't have datalad-cds
installed at the time of cloning a dataset. Thus, auto-enabling leads to
some potentially confusing error messages and then leaves the repository
in a state in which the remote is disabled. The message when cloning a
repository with a not-auto-enabling cds remote is more appropriate, I
think.